worked on medication conflicts
