+++ src/backend/.gitignore
@@ -25,4 +25,4 @@ yarn-debug.log*
 yarn-error.log*
 
 # Environment files (local configs)
-/.env.*
+/.env
