+++ src/backend/firebase/calendar_calculations.py
